+++ classes/Api/ApiRouter.php
@@ -420,6 +420,7 @@ class ApiRouter extends ProcessorBase
         $r->addRoute('POST', '/media', [MediaController::class, 'uploadSiteMedia']);
         $r->addRoute('POST', '/media/folders', [MediaController::class, 'createFolder']);
         $r->addRoute('POST', '/media/rename', [MediaController::class, 'renameFile']);
+        $r->addRoute('POST', '/media/order', [MediaController::class, 'setSiteMediaOrder']);
         $r->addRoute('POST', '/media/folders/rename', [MediaController::class, 'renameFolder']);
         $r->addRoute('DELETE', '/media/folders/{path:.+}', [MediaController::class, 'deleteFolder']);
         $r->addRoute('DELETE', '/media/{filename:.+}', [MediaController::class, 'deleteSiteMedia']);